+++ docs_that_run/__init__.py
@@ -13,4 +13,4 @@ __all__ = [
     "parse_markdown_text",
 ]
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"